Move load balancer node configuration to switches.json (#17)
* Use a json config as argument for controller

Previously, the p4info, bmv2_json and
multi_switch arguments were passed to the
controller as command line arguments and the
switch configuration was hardcoded. This commit
makes it possible to configure any number of
switches in the configuration file, and also to
customize the parameters of individual switches.

* Format all json runtime files

* Add comment to the controller Makefile target

* Add the lb_nodes field to the switches JSON config

The lb_nodes field is a list of the nodes that
the load balancer routes traffic to. This field
consists of a list of dictionaries, where each
dictionary contains the following fields:
- ip: the IP address of the node
- mac: the MAC address of the node
- port: the switch port to use to route traffic
  to the node
The controller uses this field to program the
load balancer switch with proper match-action
table entries.
The controller tracks state of the table entries
to allow updates after a container migration.