Add Redis migration example (#18)
* Use a json config as argument for controller

Previously, the p4info, bmv2_json and
multi_switch arguments were passed to the
controller as command line arguments and the
switch configuration was hardcoded. This commit
makes it possible to configure any number of
switches in the configuration file, and also to
customize the parameters of individual switches.

* Format all json runtime files

* Add comment to the controller Makefile target

* Add the lb_nodes field to the switches JSON config

The lb_nodes field is a list of the nodes that
the load balancer routes traffic to. This field
consists of a list of dictionaries, where each
dictionary contains the following fields:
- ip: the IP address of the node
- mac: the MAC address of the node
- port: the switch port to use to route traffic
  to the node
The controller uses this field to program the
load balancer switch with proper match-action
table entries.
The controller tracks state of the table entries
to allow updates after a container migration.

* Add an example for migrating a Redis container

Added two go applications - app and producer.
App acts as a backend server, providing an
endpoint to fetch data from Redis.
Producer increments a counter in Redis every
second to simulate data generation.
Both applications are containerized and can be
built using the provided Containerfiles.

* Add switches.json and s1-runtime.json

* Correct the load balancer configuration

* Include Redis uptime in the HTTP response

This is in the app server, in the /get endpoint.
The uptime is fetched from the Redis INFO command.
The uptime demonstrates successful migration.

* Add client, include connected_clients in response

The client sends requests to app and prints the
response every second.

* Update readme, add hello-world container to h3-pod

The hello-world container is added to the h3-pod
so that the h3 network is created. Otherwise, the
network would be started after restoration of the
container in h3-pod, which breaks the TCP
connection.

* Rename README to README.md

* Update the 'p4c' image name to 'docker.io/p4lang/p4c'

* Update the 'golang:1.22.2' image name to 'docker.io/golang:1.22.2'

* Drop RST packets originating from bridges

On some systems, the bridge sends a RST packet
when a TCP packet is sent to it. This commit
adds iptables rules to block RST packets
originating from h1-br, h2-br, and h3-br.

* Add sudo -E to edit_files_img.py call